+++ rln/src/circuit/iden3calc/proto.rs
@@ -1,4 +1,4 @@
-// This crate has been generated by prost-build during compilation of the code by iden3
+// This module has been generated by prost-build during compilation of the code by iden3
 // and modified manually. The *.proto file used to generate this on can be found here:
 // https://github.com/iden3/circom-witnesscalc/blob/5cb365b6e4d9052ecc69d4567fcf5bc061c20e94/protos/messages.proto
