+++ rln/src/ffi/ffi_rln.rs
@@ -1,30 +1,454 @@
 #![allow(non_camel_case_types)]
 
-use num_bigint::BigInt;
-use safer_ffi::{boxed::Box_, derive_ReprC, ffi_export, prelude::repr_c};
-#[cfg(not(feature = "stateless"))]
-use {safer_ffi::prelude::char_p, std::fs::File, std::io::Read};
+use std::{fs::File, io::Read, str::FromStr};
+
+use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
+use safer_ffi::{
+    boxed::Box_,
+    derive_ReprC, ffi_export,
+    prelude::{char_p, repr_c},
+};
+use zerokit_utils::merkle_tree::{
+    FullMerkleTree, Hasher, OptimalMerkleTree, ZerokitMerkleProof, ZerokitMerkleTree,
+};
 
 use super::ffi_utils::{CBoolResult, CFr, CResult};
 use crate::prelude::*;
 
-#[cfg(not(feature = "stateless"))]
-const MAX_CONFIG_SIZE: u64 = 1024 * 1024; // 1MB
+const MAX_CONFIG_SIZE: u64 = 1024 * 1024;
+const NO_STATELESS_TREE_ERR: &str = "tree op unsupported on stateless RLN";
+
+pub(crate) enum FFI_RLN_Inner {
+    Stateless(RLN<Stateless, ArkGroth16Backend>),
+    StatefulFullMerkleTree(RLN<Stateful<FullMerkleTree<PoseidonHash>>, ArkGroth16Backend>),
+    StatefulOptimalMerkleTree(RLN<Stateful<OptimalMerkleTree<PoseidonHash>>, ArkGroth16Backend>),
+    StatefulPmTree(RLN<Stateful<PmTree>, ArkGroth16Backend>),
+}
+
+impl FFI_RLN_Inner {
+    fn generate_proof(&self, witness: &RLNWitnessInput) -> Result<(Proof, RLNProofValues), String> {
+        match self {
+            Self::Stateless(r) => r.generate_proof(witness).map_err(|err| err.to_string()),
+            Self::StatefulFullMerkleTree(r) => {
+                r.generate_proof(witness).map_err(|err| err.to_string())
+            }
+            Self::StatefulOptimalMerkleTree(r) => {
+                r.generate_proof(witness).map_err(|err| err.to_string())
+            }
+            Self::StatefulPmTree(r) => r.generate_proof(witness).map_err(|err| err.to_string()),
+        }
+    }
+
+    fn verify(&self, proof: &Proof, values: &RLNProofValues) -> Result<bool, String> {
+        match self {
+            Self::Stateless(r) => r.verify(proof, values).map_err(|err| err.to_string()),
+            Self::StatefulFullMerkleTree(r) => {
+                r.verify(proof, values).map_err(|err| err.to_string())
+            }
+            Self::StatefulOptimalMerkleTree(r) => {
+                r.verify(proof, values).map_err(|err| err.to_string())
+            }
+            Self::StatefulPmTree(r) => r.verify(proof, values).map_err(|err| err.to_string()),
+        }
+    }
+
+    fn verify_with_roots(
+        &self,
+        proof: &Proof,
+        values: &RLNProofValues,
+        x: &Fr,
+        roots: &[Fr],
+    ) -> Result<bool, String> {
+        match self {
+            Self::Stateless(r) => r
+                .verify_with_roots(proof, values, x, roots)
+                .map_err(|err| err.to_string()),
+            Self::StatefulFullMerkleTree(r) => r
+                .verify_with_roots(proof, values, x, roots)
+                .map_err(|err| err.to_string()),
+            Self::StatefulOptimalMerkleTree(r) => r
+                .verify_with_roots(proof, values, x, roots)
+                .map_err(|err| err.to_string()),
+            Self::StatefulPmTree(r) => r
+                .verify_with_roots(proof, values, x, roots)
+                .map_err(|err| err.to_string()),
+        }
+    }
+
+    fn generate_partial_proof(
+        &self,
+        partial_witness: &RLNPartialWitnessInput,
+    ) -> Result<PartialProof, String> {
+        match self {
+            Self::Stateless(r) => r
+                .generate_partial_proof(partial_witness)
+                .map_err(|err| err.to_string()),
+            Self::StatefulFullMerkleTree(r) => r
+                .generate_partial_proof(partial_witness)
+                .map_err(|err| err.to_string()),
+            Self::StatefulOptimalMerkleTree(r) => r
+                .generate_partial_proof(partial_witness)
+                .map_err(|err| err.to_string()),
+            Self::StatefulPmTree(r) => r
+                .generate_partial_proof(partial_witness)
+                .map_err(|err| err.to_string()),
+        }
+    }
+
+    fn finish_proof(
+        &self,
+        partial_proof: &PartialProof,
+        witness: &RLNWitnessInput,
+    ) -> Result<(Proof, RLNProofValues), String> {
+        match self {
+            Self::Stateless(r) => r
+                .finish_proof(partial_proof, witness)
+                .map_err(|err| err.to_string()),
+            Self::StatefulFullMerkleTree(r) => r
+                .finish_proof(partial_proof, witness)
+                .map_err(|err| err.to_string()),
+            Self::StatefulOptimalMerkleTree(r) => r
+                .finish_proof(partial_proof, witness)
+                .map_err(|err| err.to_string()),
+            Self::StatefulPmTree(r) => r
+                .finish_proof(partial_proof, witness)
+                .map_err(|err| err.to_string()),
+        }
+    }
+
+    fn get_root(&self) -> Result<Fr, String> {
+        match self {
+            Self::Stateless(_) => Err(NO_STATELESS_TREE_ERR.to_string()),
+            Self::StatefulFullMerkleTree(r) => Ok(r.get_root()),
+            Self::StatefulOptimalMerkleTree(r) => Ok(r.get_root()),
+            Self::StatefulPmTree(r) => Ok(r.get_root()),
+        }
+    }
+
+    fn set_leaf(&mut self, index: usize, leaf: Fr) -> Result<(), String> {
+        match self {
+            Self::Stateless(_) => Err(NO_STATELESS_TREE_ERR.to_string()),
+            Self::StatefulFullMerkleTree(r) => {
+                r.set_leaf(index, leaf).map_err(|err| err.to_string())
+            }
+            Self::StatefulOptimalMerkleTree(r) => {
+                r.set_leaf(index, leaf).map_err(|err| err.to_string())
+            }
+            Self::StatefulPmTree(r) => r.set_leaf(index, leaf).map_err(|err| err.to_string()),
+        }
+    }
+
+    fn get_leaf(&self, index: usize) -> Result<Fr, String> {
+        match self {
+            Self::Stateless(_) => Err(NO_STATELESS_TREE_ERR.to_string()),
+            Self::StatefulFullMerkleTree(r) => r.get_leaf(index).map_err(|err| err.to_string()),
+            Self::StatefulOptimalMerkleTree(r) => r.get_leaf(index).map_err(|err| err.to_string()),
+            Self::StatefulPmTree(r) => r.get_leaf(index).map_err(|err| err.to_string()),
+        }
+    }
+
+    fn leaves_set(&self) -> Result<usize, String> {
+        match self {
+            Self::Stateless(_) => Err(NO_STATELESS_TREE_ERR.to_string()),
+            Self::StatefulFullMerkleTree(r) => Ok(r.leaves_set()),
+            Self::StatefulOptimalMerkleTree(r) => Ok(r.leaves_set()),
+            Self::StatefulPmTree(r) => Ok(r.leaves_set()),
+        }
+    }
+
+    fn set_next_leaf(&mut self, leaf: Fr) -> Result<(), String> {
+        match self {
+            Self::Stateless(_) => Err(NO_STATELESS_TREE_ERR.to_string()),
+            Self::StatefulFullMerkleTree(r) => r.set_next_leaf(leaf).map_err(|err| err.to_string()),
+            Self::StatefulOptimalMerkleTree(r) => {
+                r.set_next_leaf(leaf).map_err(|err| err.to_string())
+            }
+            Self::StatefulPmTree(r) => r.set_next_leaf(leaf).map_err(|err| err.to_string()),
+        }
+    }
+
+    fn set_leaves_from(&mut self, index: usize, leaves: Vec<Fr>) -> Result<(), String> {
+        match self {
+            Self::Stateless(_) => Err(NO_STATELESS_TREE_ERR.to_string()),
+            Self::StatefulFullMerkleTree(r) => r
+                .set_leaves_from(index, leaves)
+                .map_err(|err| err.to_string()),
+            Self::StatefulOptimalMerkleTree(r) => r
+                .set_leaves_from(index, leaves)
+                .map_err(|err| err.to_string()),
+            Self::StatefulPmTree(r) => r
+                .set_leaves_from(index, leaves)
+                .map_err(|err| err.to_string()),
+        }
+    }
+
+    fn delete_leaf(&mut self, index: usize) -> Result<(), String> {
+        match self {
+            Self::Stateless(_) => Err(NO_STATELESS_TREE_ERR.to_string()),
+            Self::StatefulFullMerkleTree(r) => r.delete_leaf(index).map_err(|err| err.to_string()),
+            Self::StatefulOptimalMerkleTree(r) => {
+                r.delete_leaf(index).map_err(|err| err.to_string())
+            }
+            Self::StatefulPmTree(r) => r.delete_leaf(index).map_err(|err| err.to_string()),
+        }
+    }
+
+    fn init_tree_with_leaves(&mut self, leaves: Vec<Fr>) -> Result<(), String> {
+        match self {
+            Self::Stateless(_) => Err(NO_STATELESS_TREE_ERR.to_string()),
+            Self::StatefulFullMerkleTree(r) => r
+                .init_tree_with_leaves(leaves)
+                .map_err(|err| err.to_string()),
+            Self::StatefulOptimalMerkleTree(r) => r
+                .init_tree_with_leaves(leaves)
+                .map_err(|err| err.to_string()),
+            Self::StatefulPmTree(r) => r
+                .init_tree_with_leaves(leaves)
+                .map_err(|err| err.to_string()),
+        }
+    }
+
+    fn atomic_operation(
+        &mut self,
+        index: usize,
+        leaves: Vec<Fr>,
+        indices: Vec<usize>,
+    ) -> Result<(), String> {
+        match self {
+            Self::Stateless(_) => Err(NO_STATELESS_TREE_ERR.to_string()),
+            Self::StatefulFullMerkleTree(r) => r
+                .atomic_operation(index, leaves, indices)
+                .map_err(|err| err.to_string()),
+            Self::StatefulOptimalMerkleTree(r) => r
+                .atomic_operation(index, leaves, indices)
+                .map_err(|err| err.to_string()),
+            Self::StatefulPmTree(r) => r
+                .atomic_operation(index, leaves, indices)
+                .map_err(|err| err.to_string()),
+        }
+    }
+
+    fn get_merkle_proof(&self, index: usize) -> Result<(Vec<Fr>, Vec<u8>), String> {
+        match self {
+            Self::Stateless(_) => Err(NO_STATELESS_TREE_ERR.to_string()),
+            Self::StatefulFullMerkleTree(r) => {
+                let p = r.get_merkle_proof(index).map_err(|err| err.to_string())?;
+                Ok((p.get_path_elements(), p.get_path_index()))
+            }
+            Self::StatefulOptimalMerkleTree(r) => {
+                let p = r.get_merkle_proof(index).map_err(|err| err.to_string())?;
+                Ok((p.get_path_elements(), p.get_path_index()))
+            }
+            Self::StatefulPmTree(r) => {
+                let p = r.get_merkle_proof(index).map_err(|err| err.to_string())?;
+                Ok((p.get_path_elements(), p.get_path_index()))
+            }
+        }
+    }
+
+    fn set_metadata(&mut self, metadata: &[u8]) -> Result<(), String> {
+        match self {
+            Self::Stateless(_) => Err(NO_STATELESS_TREE_ERR.to_string()),
+            Self::StatefulFullMerkleTree(r) => {
+                r.set_metadata(metadata).map_err(|err| err.to_string())
+            }
+            Self::StatefulOptimalMerkleTree(r) => {
+                r.set_metadata(metadata).map_err(|err| err.to_string())
+            }
+            Self::StatefulPmTree(r) => r.set_metadata(metadata).map_err(|err| err.to_string()),
+        }
+    }
+
+    fn get_metadata(&self) -> Result<Vec<u8>, String> {
+        match self {
+            Self::Stateless(_) => Err(NO_STATELESS_TREE_ERR.to_string()),
+            Self::StatefulFullMerkleTree(r) => r.get_metadata().map_err(|err| err.to_string()),
+            Self::StatefulOptimalMerkleTree(r) => r.get_metadata().map_err(|err| err.to_string()),
+            Self::StatefulPmTree(r) => r.get_metadata().map_err(|err| err.to_string()),
+        }
+    }
+
+    fn flush(&mut self) -> Result<(), String> {
+        match self {
+            Self::Stateless(_) => Err(NO_STATELESS_TREE_ERR.to_string()),
+            Self::StatefulFullMerkleTree(r) => r.flush().map_err(|err| err.to_string()),
+            Self::StatefulOptimalMerkleTree(r) => r.flush().map_err(|err| err.to_string()),
+            Self::StatefulPmTree(r) => r.flush().map_err(|err| err.to_string()),
+        }
+    }
+}
+
+impl From<RLN<Stateless, ArkGroth16Backend>> for FFI_RLN_Inner {
+    fn from(r: RLN<Stateless, ArkGroth16Backend>) -> Self {
+        Self::Stateless(r)
+    }
+}
+
+impl From<RLN<Stateful<FullMerkleTree<PoseidonHash>>, ArkGroth16Backend>> for FFI_RLN_Inner {
+    fn from(r: RLN<Stateful<FullMerkleTree<PoseidonHash>>, ArkGroth16Backend>) -> Self {
+        Self::StatefulFullMerkleTree(r)
+    }
+}
+
+impl From<RLN<Stateful<OptimalMerkleTree<PoseidonHash>>, ArkGroth16Backend>> for FFI_RLN_Inner {
+    fn from(r: RLN<Stateful<OptimalMerkleTree<PoseidonHash>>, ArkGroth16Backend>) -> Self {
+        Self::StatefulOptimalMerkleTree(r)
+    }
+}
+
+impl From<RLN<Stateful<PmTree>, ArkGroth16Backend>> for FFI_RLN_Inner {
+    fn from(r: RLN<Stateful<PmTree>, ArkGroth16Backend>) -> Self {
+        Self::StatefulPmTree(r)
+    }
+}
 
 // FFI_RLN
 
 #[derive_ReprC]
 #[repr(opaque)]
-pub struct FFI_RLN(pub(crate) RLN);
+pub struct FFI_RLN(pub(crate) FFI_RLN_Inner);
 
-// RLN initialization APIs
+fn parse_zkey_and_graph(
+    zkey_data: &repr_c::Vec<u8>,
+    graph_data: &repr_c::Vec<u8>,
+) -> Result<(Zkey, Graph), String> {
+    let zkey = zkey_from_raw(zkey_data).map_err(|err| err.to_string())?;
+    let graph = graph_from_raw(graph_data, None, None).map_err(|err| err.to_string())?;
+    Ok((zkey, graph))
+}
 
-#[cfg(not(feature = "stateless"))]
 #[ffi_export]
-pub fn ffi_rln_new(
+pub fn ffi_rln_new_stateless_default() -> repr_c::Box<FFI_RLN> {
+    let rln = RLNBuilder::stateless().build();
+    Box_::new(FFI_RLN(rln.into()))
+}
+
+#[ffi_export]
+pub fn ffi_rln_new_stateless(
+    zkey_data: &repr_c::Vec<u8>,
+    graph_data: &repr_c::Vec<u8>,
+) -> CResult<repr_c::Box<FFI_RLN>, repr_c::String> {
+    match parse_zkey_and_graph(zkey_data, graph_data) {
+        Ok((zkey, graph)) => {
+            let rln = RLNBuilder::stateless().graph(graph).zkey(zkey).build();
+            CResult {
+                ok: Some(Box_::new(FFI_RLN(rln.into()))),
+                err: None,
+            }
+        }
+        Err(err) => CResult {
+            ok: None,
+            err: Some(err.into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_rln_new_with_full_merkle_tree_default() -> repr_c::Box<FFI_RLN> {
+    let full_merkle_tree = FullMerkleTree::<PoseidonHash>::default(DEFAULT_TREE_DEPTH).unwrap();
+    let rln = RLNBuilder::stateful().tree(full_merkle_tree).build();
+    Box_::new(FFI_RLN(rln.into()))
+}
+
+#[ffi_export]
+pub fn ffi_rln_new_with_full_merkle_tree(
     tree_depth: usize,
+    zkey_data: &repr_c::Vec<u8>,
+    graph_data: &repr_c::Vec<u8>,
+) -> CResult<repr_c::Box<FFI_RLN>, repr_c::String> {
+    let (zkey, graph) = match parse_zkey_and_graph(zkey_data, graph_data) {
+        Ok(parsed) => parsed,
+        Err(err) => {
+            return CResult {
+                ok: None,
+                err: Some(err.into()),
+            }
+        }
+    };
+    match FullMerkleTree::<PoseidonHash>::default(tree_depth) {
+        Ok(full_merkle_tree) => {
+            let rln = RLNBuilder::stateful()
+                .tree(full_merkle_tree)
+                .graph(graph)
+                .zkey(zkey)
+                .build();
+            CResult {
+                ok: Some(Box_::new(FFI_RLN(rln.into()))),
+                err: None,
+            }
+        }
+        Err(err) => CResult {
+            ok: None,
+            err: Some(err.to_string().into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_rln_new_with_optimal_merkle_tree_default() -> repr_c::Box<FFI_RLN> {
+    let optimal_merkle_tree =
+        OptimalMerkleTree::<PoseidonHash>::default(DEFAULT_TREE_DEPTH).unwrap();
+    let rln = RLNBuilder::stateful().tree(optimal_merkle_tree).build();
+    Box_::new(FFI_RLN(rln.into()))
+}
+
+#[ffi_export]
+pub fn ffi_rln_new_with_optimal_merkle_tree(
+    tree_depth: usize,
+    zkey_data: &repr_c::Vec<u8>,
+    graph_data: &repr_c::Vec<u8>,
+) -> CResult<repr_c::Box<FFI_RLN>, repr_c::String> {
+    let (zkey, graph) = match parse_zkey_and_graph(zkey_data, graph_data) {
+        Ok(parsed) => parsed,
+        Err(err) => {
+            return CResult {
+                ok: None,
+                err: Some(err.into()),
+            }
+        }
+    };
+    match OptimalMerkleTree::<PoseidonHash>::default(tree_depth) {
+        Ok(optimal_merkle_tree) => {
+            let rln = RLNBuilder::stateful()
+                .tree(optimal_merkle_tree)
+                .graph(graph)
+                .zkey(zkey)
+                .build();
+            CResult {
+                ok: Some(Box_::new(FFI_RLN(rln.into()))),
+                err: None,
+            }
+        }
+        Err(err) => CResult {
+            ok: None,
+            err: Some(err.to_string().into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_rln_new_with_pm_tree_default() -> repr_c::Box<FFI_RLN> {
+    let pm_tree = PmTree::default(DEFAULT_TREE_DEPTH).unwrap();
+    let rln = RLNBuilder::stateful().tree(pm_tree).build();
+    Box_::new(FFI_RLN(rln.into()))
+}
+
+#[ffi_export]
+pub fn ffi_rln_new_with_pm_tree(
+    tree_depth: usize,
+    zkey_data: &repr_c::Vec<u8>,
+    graph_data: &repr_c::Vec<u8>,
     config_path: char_p::Ref<'_>,
 ) -> CResult<repr_c::Box<FFI_RLN>, repr_c::String> {
+    let (zkey, graph) = match parse_zkey_and_graph(zkey_data, graph_data) {
+        Ok(parsed) => parsed,
+        Err(err) => {
+            return CResult {
+                ok: None,
+                err: Some(err.into()),
+            }
+        }
+    };
     let config_str = File::open(config_path.to_str())
         .and_then(|mut file| {
             let metadata = file.metadata()?;
@@ -43,71 +467,32 @@ pub fn ffi_rln_new(
             Ok(s)
         })
         .unwrap_or_default();
-
-    match RLN::new(tree_depth, config_str.as_str()) {
-        Ok(rln) => CResult {
-            ok: Some(Box_::new(FFI_RLN(rln))),
-            err: None,
-        },
-        Err(err) => CResult {
-            ok: None,
-            err: Some(err.to_string().into()),
-        },
-    }
-}
-
-#[cfg(feature = "stateless")]
-#[ffi_export]
-pub fn ffi_rln_new() -> CResult<repr_c::Box<FFI_RLN>, repr_c::String> {
-    match RLN::new() {
-        Ok(rln) => CResult {
-            ok: Some(Box_::new(FFI_RLN(rln))),
-            err: None,
-        },
-        Err(err) => CResult {
-            ok: None,
-            err: Some(err.to_string().into()),
-        },
-    }
-}
-
-#[cfg(not(feature = "stateless"))]
-#[ffi_export]
-pub fn ffi_rln_new_with_params(
-    tree_depth: usize,
-    zkey_data: &repr_c::Vec<u8>,
-    graph_data: &repr_c::Vec<u8>,
-    config_path: char_p::Ref<'_>,
-) -> CResult<repr_c::Box<FFI_RLN>, repr_c::String> {
-    let config_str = File::open(config_path.to_str())
-        .and_then(|mut file| {
-            let metadata = file.metadata()?;
-            if metadata.len() > MAX_CONFIG_SIZE {
-                return Err(std::io::Error::new(
-                    std::io::ErrorKind::InvalidData,
-                    format!(
-                        "Config file too large: {} bytes (max {} bytes)",
-                        metadata.len(),
-                        MAX_CONFIG_SIZE
-                    ),
-                ));
+    let pm_tree = if config_str.is_empty() {
+        PmTree::default(tree_depth)
+    } else {
+        let cfg = match PmTreeConfig::from_str(&config_str) {
+            Ok(c) => c,
+            Err(err) => {
+                return CResult {
+                    ok: None,
+                    err: Some(err.to_string().into()),
+                }
             }
-            let mut s = String::new();
-            file.read_to_string(&mut s)?;
-            Ok(s)
-        })
-        .unwrap_or_default();
-
-    match RLN::new_with_params(
-        tree_depth,
-        zkey_data.to_vec(),
-        graph_data.to_vec(),
-        config_str.as_str(),
-    ) {
-        Ok(rln) => CResult {
-            ok: Some(Box_::new(FFI_RLN(rln))),
-            err: None,
-        },
+        };
+        PmTree::new(tree_depth, PoseidonHash::default_leaf(), cfg)
+    };
+    match pm_tree {
+        Ok(pm_tree) => {
+            let rln = RLNBuilder::stateful()
+                .tree(pm_tree)
+                .graph(graph)
+                .zkey(zkey)
+                .build();
+            CResult {
+                ok: Some(Box_::new(FFI_RLN(rln.into()))),
+                err: None,
+            }
+        }
         Err(err) => CResult {
             ok: None,
             err: Some(err.to_string().into()),
@@ -115,20 +500,101 @@ pub fn ffi_rln_new_with_params(
     }
 }
 
-#[cfg(feature = "stateless")]
 #[ffi_export]
-pub fn ffi_rln_new_with_params(
-    zkey_data: &repr_c::Vec<u8>,
-    graph_data: &repr_c::Vec<u8>,
-) -> CResult<repr_c::Box<FFI_RLN>, repr_c::String> {
-    match RLN::new_with_params(zkey_data.to_vec(), graph_data.to_vec()) {
-        Ok(rln) => CResult {
-            ok: Some(Box_::new(FFI_RLN(rln))),
+pub fn ffi_rln_generate_proof(
+    rln: &repr_c::Box<FFI_RLN>,
+    witness: &repr_c::Box<FFI_RLNWitnessInput>,
+) -> CResult<repr_c::Box<FFI_RLNProof>, repr_c::String> {
+    match rln.0.generate_proof(&witness.0) {
+        Ok((proof, values)) => CResult {
+            ok: Some(Box_::new(FFI_RLNProof(RLNProof::new(proof, values)))),
             err: None,
         },
         Err(err) => CResult {
             ok: None,
-            err: Some(err.to_string().into()),
+            err: Some(err.into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_rln_verify(
+    rln: &repr_c::Box<FFI_RLN>,
+    rln_proof: &repr_c::Box<FFI_RLNProof>,
+    x: &CFr,
+) -> CBoolResult {
+    if rln_proof.0.values.x() != x.0 {
+        return CBoolResult {
+            ok: false,
+            err: None,
+        };
+    }
+    match rln.0.verify(&rln_proof.0.proof, &rln_proof.0.values) {
+        Ok(verified) => CBoolResult {
+            ok: verified,
+            err: None,
+        },
+        Err(err) => CBoolResult {
+            ok: false,
+            err: Some(err.into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_rln_verify_with_roots(
+    rln: &repr_c::Box<FFI_RLN>,
+    rln_proof: &repr_c::Box<FFI_RLNProof>,
+    roots: &repr_c::Vec<CFr>,
+    x: &CFr,
+) -> CBoolResult {
+    let roots_fr: Vec<Fr> = roots.iter().map(|cfr| cfr.0).collect();
+    match rln
+        .0
+        .verify_with_roots(&rln_proof.0.proof, &rln_proof.0.values, &x.0, &roots_fr)
+    {
+        Ok(verified) => CBoolResult {
+            ok: verified,
+            err: None,
+        },
+        Err(err) => CBoolResult {
+            ok: false,
+            err: Some(err.into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_rln_generate_partial_proof(
+    rln: &repr_c::Box<FFI_RLN>,
+    partial_witness: &repr_c::Box<FFI_RLNPartialWitnessInput>,
+) -> CResult<repr_c::Box<FFI_RLNPartialProof>, repr_c::String> {
+    match rln.0.generate_partial_proof(&partial_witness.0) {
+        Ok(pp) => CResult {
+            ok: Some(Box_::new(FFI_RLNPartialProof(pp))),
+            err: None,
+        },
+        Err(err) => CResult {
+            ok: None,
+            err: Some(err.into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_rln_finish_proof(
+    rln: &repr_c::Box<FFI_RLN>,
+    partial_proof: &repr_c::Box<FFI_RLNPartialProof>,
+    witness: &repr_c::Box<FFI_RLNWitnessInput>,
+) -> CResult<repr_c::Box<FFI_RLNProof>, repr_c::String> {
+    match rln.0.finish_proof(&partial_proof.0, &witness.0) {
+        Ok((proof, values)) => CResult {
+            ok: Some(Box_::new(FFI_RLNProof(RLNProof::new(proof, values)))),
+            err: None,
+        },
+        Err(err) => CResult {
+            ok: None,
+            err: Some(err.into()),
         },
     }
 }
@@ -138,186 +604,7 @@ pub fn ffi_rln_free(rln: repr_c::Box<FFI_RLN>) {
     drop(rln);
 }
 
-#[ffi_export]
-pub fn ffi_rln_get_tree_depth(rln: &repr_c::Box<FFI_RLN>) -> usize {
-    rln.0.tree_depth()
-}
-
-#[ffi_export]
-pub fn ffi_rln_get_max_out(rln: &repr_c::Box<FFI_RLN>) -> usize {
-    rln.0.max_out()
-}
-
-// RLNProof
-
-#[derive_ReprC]
-#[repr(opaque)]
-pub struct FFI_RLNProof(pub(crate) RLNProof);
-
-#[ffi_export]
-pub fn ffi_rln_proof_get_values(
-    rln_proof: &repr_c::Box<FFI_RLNProof>,
-) -> repr_c::Box<FFI_RLNProofValues> {
-    Box_::new(FFI_RLNProofValues(rln_proof.0.proof_values.clone()))
-}
-
-#[ffi_export]
-pub fn ffi_rln_proof_get_version_byte(rln_proof: &repr_c::Box<FFI_RLNProof>) -> u8 {
-    rln_proof.0.version_byte()
-}
-
-#[ffi_export]
-pub fn ffi_rln_proof_to_bytes_le(
-    rln_proof: &repr_c::Box<FFI_RLNProof>,
-) -> CResult<repr_c::Vec<u8>, repr_c::String> {
-    match rln_proof_to_bytes_le(&rln_proof.0) {
-        Ok(bytes) => CResult {
-            ok: Some(bytes.into()),
-            err: None,
-        },
-        Err(err) => CResult {
-            ok: None,
-            err: Some(err.to_string().into()),
-        },
-    }
-}
-
-#[ffi_export]
-pub fn ffi_rln_proof_to_bytes_be(
-    rln_proof: &repr_c::Box<FFI_RLNProof>,
-) -> CResult<repr_c::Vec<u8>, repr_c::String> {
-    match rln_proof_to_bytes_be(&rln_proof.0) {
-        Ok(bytes) => CResult {
-            ok: Some(bytes.into()),
-            err: None,
-        },
-        Err(err) => CResult {
-            ok: None,
-            err: Some(err.to_string().into()),
-        },
-    }
-}
-
-#[ffi_export]
-pub fn ffi_bytes_le_to_rln_proof(
-    bytes: &repr_c::Vec<u8>,
-) -> CResult<repr_c::Box<FFI_RLNProof>, repr_c::String> {
-    match bytes_le_to_rln_proof(bytes) {
-        Ok((rln_proof, _)) => CResult {
-            ok: Some(Box_::new(FFI_RLNProof(rln_proof))),
-            err: None,
-        },
-        Err(err) => CResult {
-            ok: None,
-            err: Some(err.to_string().into()),
-        },
-    }
-}
-
-#[ffi_export]
-pub fn ffi_bytes_be_to_rln_proof(
-    bytes: &repr_c::Vec<u8>,
-) -> CResult<repr_c::Box<FFI_RLNProof>, repr_c::String> {
-    match bytes_be_to_rln_proof(bytes) {
-        Ok((rln_proof, _)) => CResult {
-            ok: Some(Box_::new(FFI_RLNProof(rln_proof))),
-            err: None,
-        },
-        Err(err) => CResult {
-            ok: None,
-            err: Some(err.to_string().into()),
-        },
-    }
-}
-
-#[ffi_export]
-pub fn ffi_rln_proof_free(rln_proof: repr_c::Box<FFI_RLNProof>) {
-    drop(rln_proof);
-}
-
-// RLNPartialProof
-
-#[derive_ReprC]
-#[repr(opaque)]
-pub struct FFI_RLNPartialProof(pub(crate) PartialProof);
-
-#[ffi_export]
-pub fn ffi_rln_partial_proof_get_version_byte(
-    partial_proof: &repr_c::Box<FFI_RLNPartialProof>,
-) -> u8 {
-    partial_proof.0.version_byte()
-}
-
-#[ffi_export]
-pub fn ffi_rln_partial_proof_to_bytes_le(
-    partial_proof: &repr_c::Box<FFI_RLNPartialProof>,
-) -> CResult<repr_c::Vec<u8>, repr_c::String> {
-    match rln_partial_proof_to_bytes_le(&partial_proof.0) {
-        Ok(bytes) => CResult {
-            ok: Some(bytes.into()),
-            err: None,
-        },
-        Err(err) => CResult {
-            ok: None,
-            err: Some(err.to_string().into()),
-        },
-    }
-}
-
-#[ffi_export]
-pub fn ffi_bytes_le_to_rln_partial_proof(
-    bytes: &repr_c::Vec<u8>,
-) -> CResult<repr_c::Box<FFI_RLNPartialProof>, repr_c::String> {
-    match bytes_le_to_rln_partial_proof(bytes) {
-        Ok((partial_proof, _)) => CResult {
-            ok: Some(Box_::new(FFI_RLNPartialProof(partial_proof))),
-            err: None,
-        },
-        Err(err) => CResult {
-            ok: None,
-            err: Some(err.to_string().into()),
-        },
-    }
-}
-
-#[ffi_export]
-pub fn ffi_rln_partial_proof_free(partial_proof: repr_c::Box<FFI_RLNPartialProof>) {
-    drop(partial_proof);
-}
-
-#[ffi_export]
-pub fn ffi_rln_partial_proof_to_bytes_be(
-    partial_proof: &repr_c::Box<FFI_RLNPartialProof>,
-) -> CResult<repr_c::Vec<u8>, repr_c::String> {
-    match rln_partial_proof_to_bytes_be(&partial_proof.0) {
-        Ok(bytes) => CResult {
-            ok: Some(bytes.into()),
-            err: None,
-        },
-        Err(err) => CResult {
-            ok: None,
-            err: Some(err.to_string().into()),
-        },
-    }
-}
-
-#[ffi_export]
-pub fn ffi_bytes_be_to_rln_partial_proof(
-    bytes: &repr_c::Vec<u8>,
-) -> CResult<repr_c::Box<FFI_RLNPartialProof>, repr_c::String> {
-    match bytes_be_to_rln_partial_proof(bytes) {
-        Ok((partial_proof, _)) => CResult {
-            ok: Some(Box_::new(FFI_RLNPartialProof(partial_proof))),
-            err: None,
-        },
-        Err(err) => CResult {
-            ok: None,
-            err: Some(err.to_string().into()),
-        },
-    }
-}
-
-// RLNWitnessInput
+// FFI_RLNWitnessInput
 
 #[derive_ReprC]
 #[repr(opaque)]
@@ -337,17 +624,18 @@ pub fn ffi_rln_witness_input_new_single(
     let path_elements: Vec<Fr> = path_elements.iter().map(|cfr| cfr.0).collect();
     let identity_path_index: Vec<u8> = identity_path_index.iter().copied().collect();
 
-    match RLNWitnessInput::new_single(
-        IdSecret::from(&mut identity_secret_fr),
-        user_message_limit.0,
-        message_id.0,
-        path_elements,
-        identity_path_index,
-        x.0,
-        external_nullifier.0,
-    ) {
-        Ok(witness) => CResult {
-            ok: Some(Box_::new(FFI_RLNWitnessInput(witness))),
+    match RLNWitnessInput::new_single()
+        .identity_secret(IdSecret::from(&mut identity_secret_fr))
+        .user_message_limit(user_message_limit.0)
+        .path_elements(path_elements)
+        .identity_path_index(identity_path_index)
+        .x(x.0)
+        .external_nullifier(external_nullifier.0)
+        .message_id(message_id.0)
+        .build()
+    {
+        Ok(w) => CResult {
+            ok: Some(Box_::new(FFI_RLNWitnessInput(w))),
             err: None,
         },
         Err(err) => CResult {
@@ -374,18 +662,19 @@ pub fn ffi_rln_witness_input_new_multi(
     let message_ids: Vec<Fr> = message_ids.iter().map(|cfr| cfr.0).collect();
     let selector_used: Vec<bool> = selector_used.iter().copied().collect();
 
-    match RLNWitnessInput::new_multi(
-        IdSecret::from(&mut identity_secret_fr),
-        user_message_limit.0,
-        message_ids,
-        path_elements,
-        identity_path_index,
-        x.0,
-        external_nullifier.0,
-        selector_used,
-    ) {
-        Ok(witness) => CResult {
-            ok: Some(Box_::new(FFI_RLNWitnessInput(witness))),
+    match RLNWitnessInput::new_multi()
+        .identity_secret(IdSecret::from(&mut identity_secret_fr))
+        .user_message_limit(user_message_limit.0)
+        .path_elements(path_elements)
+        .identity_path_index(identity_path_index)
+        .x(x.0)
+        .external_nullifier(external_nullifier.0)
+        .message_ids(message_ids)
+        .selector_used(selector_used)
+        .build()
+    {
+        Ok(w) => CResult {
+            ok: Some(Box_::new(FFI_RLNWitnessInput(w))),
             err: None,
         },
         Err(err) => CResult {
@@ -393,11 +682,6 @@ pub fn ffi_rln_witness_input_new_multi(
             err: Some(err.to_string().into()),
         },
     }
-}
-
-#[ffi_export]
-pub fn ffi_rln_witness_input_get_version_byte(witness: &repr_c::Box<FFI_RLNWitnessInput>) -> u8 {
-    witness.0.version_byte()
 }
 
 #[ffi_export]
@@ -411,27 +695,44 @@ pub fn ffi_rln_witness_input_get_identity_secret(
 pub fn ffi_rln_witness_input_get_user_message_limit(
     witness: &repr_c::Box<FFI_RLNWitnessInput>,
 ) -> repr_c::Box<CFr> {
-    CFr::from(*witness.0.user_message_limit()).into()
+    CFr::from(witness.0.user_message_limit()).into()
 }
 
 #[ffi_export]
 pub fn ffi_rln_witness_input_get_message_id(
     witness: &repr_c::Box<FFI_RLNWitnessInput>,
-) -> repr_c::Box<CFr> {
-    CFr::from(*witness.0.message_id()).into()
+) -> CResult<repr_c::Box<CFr>, repr_c::String> {
+    match witness.0.message_id() {
+        Some(id) => CResult {
+            ok: Some(CFr::from(id).into()),
+            err: None,
+        },
+        None => CResult {
+            ok: None,
+            err: Some("witness is Multi; use get_message_ids".into()),
+        },
+    }
 }
 
 #[ffi_export]
 pub fn ffi_rln_witness_input_get_message_ids(
     witness: &repr_c::Box<FFI_RLNWitnessInput>,
-) -> repr_c::Vec<CFr> {
-    witness
-        .0
-        .message_ids()
-        .iter()
-        .map(|fr| CFr::from(*fr))
-        .collect::<Vec<_>>()
-        .into()
+) -> CResult<repr_c::Vec<CFr>, repr_c::String> {
+    match witness.0.message_ids() {
+        Some(ids) => CResult {
+            ok: Some(
+                ids.iter()
+                    .map(|fr| CFr::from(*fr))
+                    .collect::<Vec<_>>()
+                    .into(),
+            ),
+            err: None,
+        },
+        None => CResult {
+            ok: None,
+            err: Some("witness is Single; use get_message_id".into()),
+        },
+    }
 }
 
 #[ffi_export]
@@ -456,29 +757,39 @@ pub fn ffi_rln_witness_input_get_identity_path_index(
 
 #[ffi_export]
 pub fn ffi_rln_witness_input_get_x(witness: &repr_c::Box<FFI_RLNWitnessInput>) -> repr_c::Box<CFr> {
-    CFr::from(*witness.0.x()).into()
+    CFr::from(witness.0.x()).into()
 }
 
 #[ffi_export]
 pub fn ffi_rln_witness_input_get_external_nullifier(
     witness: &repr_c::Box<FFI_RLNWitnessInput>,
 ) -> repr_c::Box<CFr> {
-    CFr::from(*witness.0.external_nullifier()).into()
+    CFr::from(witness.0.external_nullifier()).into()
 }
 
 #[ffi_export]
 pub fn ffi_rln_witness_input_get_selector_used(
     witness: &repr_c::Box<FFI_RLNWitnessInput>,
-) -> repr_c::Vec<bool> {
-    witness.0.selector_used().to_vec().into()
+) -> CResult<repr_c::Vec<bool>, repr_c::String> {
+    match witness.0.selector_used() {
+        Some(s) => CResult {
+            ok: Some(s.to_vec().into()),
+            err: None,
+        },
+        None => CResult {
+            ok: None,
+            err: Some("witness is Single; selector_used is Multi-only".into()),
+        },
+    }
 }
 
 #[ffi_export]
 pub fn ffi_rln_witness_to_bytes_le(
     witness: &repr_c::Box<FFI_RLNWitnessInput>,
 ) -> CResult<repr_c::Vec<u8>, repr_c::String> {
-    match rln_witness_to_bytes_le(&witness.0) {
-        Ok(bytes) => CResult {
+    let mut bytes = Vec::new();
+    match witness.0.serialize_compressed(&mut bytes) {
+        Ok(()) => CResult {
             ok: Some(bytes.into()),
             err: None,
         },
@@ -493,8 +804,9 @@ pub fn ffi_rln_witness_to_bytes_le(
 pub fn ffi_rln_witness_to_bytes_be(
     witness: &repr_c::Box<FFI_RLNWitnessInput>,
 ) -> CResult<repr_c::Vec<u8>, repr_c::String> {
-    match rln_witness_to_bytes_be(&witness.0) {
-        Ok(bytes) => CResult {
+    let mut bytes = Vec::new();
+    match CanonicalSerializeBE::serialize(&witness.0, &mut bytes) {
+        Ok(()) => CResult {
             ok: Some(bytes.into()),
             err: None,
         },
@@ -509,9 +821,9 @@ pub fn ffi_rln_witness_to_bytes_be(
 pub fn ffi_bytes_le_to_rln_witness(
     bytes: &repr_c::Vec<u8>,
 ) -> CResult<repr_c::Box<FFI_RLNWitnessInput>, repr_c::String> {
-    match bytes_le_to_rln_witness(bytes) {
-        Ok((witness, _)) => CResult {
-            ok: Some(Box_::new(FFI_RLNWitnessInput(witness))),
+    match RLNWitnessInput::deserialize_compressed(&bytes[..]) {
+        Ok(w) => CResult {
+            ok: Some(Box_::new(FFI_RLNWitnessInput(w))),
             err: None,
         },
         Err(err) => CResult {
@@ -525,25 +837,9 @@ pub fn ffi_bytes_le_to_rln_witness(
 pub fn ffi_bytes_be_to_rln_witness(
     bytes: &repr_c::Vec<u8>,
 ) -> CResult<repr_c::Box<FFI_RLNWitnessInput>, repr_c::String> {
-    match bytes_be_to_rln_witness(bytes) {
-        Ok((witness, _)) => CResult {
-            ok: Some(Box_::new(FFI_RLNWitnessInput(witness))),
-            err: None,
-        },
-        Err(err) => CResult {
-            ok: None,
-            err: Some(err.to_string().into()),
-        },
-    }
-}
-
-#[ffi_export]
-pub fn ffi_rln_witness_to_bigint_json(
-    witness: &repr_c::Box<FFI_RLNWitnessInput>,
-) -> CResult<repr_c::String, repr_c::String> {
-    match rln_witness_to_bigint_json(&witness.0) {
-        Ok(json) => CResult {
-            ok: Some(json.to_string().into()),
+    match <RLNWitnessInput as CanonicalDeserializeBE>::deserialize(&bytes[..]) {
+        Ok(w) => CResult {
+            ok: Some(Box_::new(FFI_RLNWitnessInput(w))),
             err: None,
         },
         Err(err) => CResult {
@@ -558,7 +854,7 @@ pub fn ffi_rln_witness_input_free(witness: repr_c::Box<FFI_RLNWitnessInput>) {
     drop(witness);
 }
 
-// RLNPartialWitnessInput
+// FFI_RLNPartialWitnessInput
 
 #[derive_ReprC]
 #[repr(opaque)]
@@ -574,14 +870,15 @@ pub fn ffi_rln_partial_witness_input_new(
     let mut identity_secret_fr = identity_secret.0;
     let path_elements: Vec<Fr> = path_elements.iter().map(|cfr| cfr.0).collect();
     let identity_path_index: Vec<u8> = identity_path_index.iter().copied().collect();
-    match RLNPartialWitnessInput::new(
-        IdSecret::from(&mut identity_secret_fr),
-        user_message_limit.0,
-        path_elements,
-        identity_path_index,
-    ) {
-        Ok(partial_witness) => CResult {
-            ok: Some(Box_::new(FFI_RLNPartialWitnessInput(partial_witness))),
+    match RLNPartialWitnessInput::new()
+        .identity_secret(IdSecret::from(&mut identity_secret_fr))
+        .user_message_limit(user_message_limit.0)
+        .path_elements(path_elements)
+        .identity_path_index(identity_path_index)
+        .build()
+    {
+        Ok(w) => CResult {
+            ok: Some(Box_::new(FFI_RLNPartialWitnessInput(w))),
             err: None,
         },
         Err(err) => CResult {
@@ -592,24 +889,17 @@ pub fn ffi_rln_partial_witness_input_new(
 }
 
 #[ffi_export]
-pub fn ffi_rln_partial_witness_input_get_version_byte(
-    witness: &repr_c::Box<FFI_RLNPartialWitnessInput>,
-) -> u8 {
-    witness.0.version_byte()
-}
-
-#[ffi_export]
 pub fn ffi_rln_partial_witness_input_get_identity_secret(
     witness: &repr_c::Box<FFI_RLNPartialWitnessInput>,
 ) -> repr_c::Box<CFr> {
-    CFr::from(**witness.0.identity_secret()).into()
+    CFr::from(*witness.0.identity_secret).into()
 }
 
 #[ffi_export]
 pub fn ffi_rln_partial_witness_input_get_user_message_limit(
     witness: &repr_c::Box<FFI_RLNPartialWitnessInput>,
 ) -> repr_c::Box<CFr> {
-    CFr::from(*witness.0.user_message_limit()).into()
+    CFr::from(witness.0.user_message_limit).into()
 }
 
 #[ffi_export]
@@ -618,7 +908,7 @@ pub fn ffi_rln_partial_witness_input_get_path_elements(
 ) -> repr_c::Vec<CFr> {
     witness
         .0
-        .path_elements()
+        .path_elements
         .iter()
         .map(|fr| CFr::from(*fr))
         .collect::<Vec<_>>()
@@ -629,7 +919,7 @@ pub fn ffi_rln_partial_witness_input_get_path_elements(
 pub fn ffi_rln_partial_witness_input_get_identity_path_index(
     witness: &repr_c::Box<FFI_RLNPartialWitnessInput>,
 ) -> repr_c::Vec<u8> {
-    witness.0.identity_path_index().to_vec().into()
+    witness.0.identity_path_index.to_vec().into()
 }
 
 #[ffi_export]
@@ -644,8 +934,9 @@ pub fn ffi_rln_witness_to_partial_witness(
 pub fn ffi_rln_partial_witness_to_bytes_le(
     witness: &repr_c::Box<FFI_RLNPartialWitnessInput>,
 ) -> CResult<repr_c::Vec<u8>, repr_c::String> {
-    match rln_partial_witness_to_bytes_le(&witness.0) {
-        Ok(bytes) => CResult {
+    let mut bytes = Vec::new();
+    match witness.0.serialize_compressed(&mut bytes) {
+        Ok(()) => CResult {
             ok: Some(bytes.into()),
             err: None,
         },
@@ -660,8 +951,9 @@ pub fn ffi_rln_partial_witness_to_bytes_le(
 pub fn ffi_rln_partial_witness_to_bytes_be(
     witness: &repr_c::Box<FFI_RLNPartialWitnessInput>,
 ) -> CResult<repr_c::Vec<u8>, repr_c::String> {
-    match rln_partial_witness_to_bytes_be(&witness.0) {
-        Ok(bytes) => CResult {
+    let mut bytes = Vec::new();
+    match CanonicalSerializeBE::serialize(&witness.0, &mut bytes) {
+        Ok(()) => CResult {
             ok: Some(bytes.into()),
             err: None,
         },
@@ -676,9 +968,9 @@ pub fn ffi_rln_partial_witness_to_bytes_be(
 pub fn ffi_bytes_le_to_rln_partial_witness(
     bytes: &repr_c::Vec<u8>,
 ) -> CResult<repr_c::Box<FFI_RLNPartialWitnessInput>, repr_c::String> {
-    match bytes_le_to_rln_partial_witness(bytes) {
-        Ok((witness, _)) => CResult {
-            ok: Some(Box_::new(FFI_RLNPartialWitnessInput(witness))),
+    match RLNPartialWitnessInput::deserialize_compressed(&bytes[..]) {
+        Ok(w) => CResult {
+            ok: Some(Box_::new(FFI_RLNPartialWitnessInput(w))),
             err: None,
         },
         Err(err) => CResult {
@@ -692,9 +984,9 @@ pub fn ffi_bytes_le_to_rln_partial_witness(
 pub fn ffi_bytes_be_to_rln_partial_witness(
     bytes: &repr_c::Vec<u8>,
 ) -> CResult<repr_c::Box<FFI_RLNPartialWitnessInput>, repr_c::String> {
-    match bytes_be_to_rln_partial_witness(bytes) {
-        Ok((witness, _)) => CResult {
-            ok: Some(Box_::new(FFI_RLNPartialWitnessInput(witness))),
+    match <RLNPartialWitnessInput as CanonicalDeserializeBE>::deserialize(&bytes[..]) {
+        Ok(w) => CResult {
+            ok: Some(Box_::new(FFI_RLNPartialWitnessInput(w))),
             err: None,
         },
         Err(err) => CResult {
@@ -709,7 +1001,135 @@ pub fn ffi_rln_partial_witness_input_free(witness: repr_c::Box<FFI_RLNPartialWit
     drop(witness);
 }
 
-// RLNProofValues
+// FFI_RLNProof
+
+#[derive_ReprC]
+#[repr(opaque)]
+pub struct FFI_RLNProof(pub(crate) RLNProof);
+
+#[ffi_export]
+pub fn ffi_rln_proof_get_values(
+    rln_proof: &repr_c::Box<FFI_RLNProof>,
+) -> repr_c::Box<FFI_RLNProofValues> {
+    Box_::new(FFI_RLNProofValues(rln_proof.0.values.clone()))
+}
+
+#[ffi_export]
+pub fn ffi_rln_proof_to_bytes_le(
+    rln_proof: &repr_c::Box<FFI_RLNProof>,
+) -> CResult<repr_c::Vec<u8>, repr_c::String> {
+    let mut bytes = Vec::new();
+    match rln_proof.0.serialize_compressed(&mut bytes) {
+        Ok(()) => CResult {
+            ok: Some(bytes.into()),
+            err: None,
+        },
+        Err(err) => CResult {
+            ok: None,
+            err: Some(err.to_string().into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_rln_proof_to_bytes_mixed(
+    rln_proof: &repr_c::Box<FFI_RLNProof>,
+) -> CResult<repr_c::Vec<u8>, repr_c::String> {
+    let mut bytes = Vec::new();
+    match CanonicalSerializeMixed::serialize(&rln_proof.0, &mut bytes) {
+        Ok(()) => CResult {
+            ok: Some(bytes.into()),
+            err: None,
+        },
+        Err(err) => CResult {
+            ok: None,
+            err: Some(err.to_string().into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_bytes_le_to_rln_proof(
+    bytes: &repr_c::Vec<u8>,
+) -> CResult<repr_c::Box<FFI_RLNProof>, repr_c::String> {
+    match RLNProof::deserialize_compressed(&bytes[..]) {
+        Ok(p) => CResult {
+            ok: Some(Box_::new(FFI_RLNProof(p))),
+            err: None,
+        },
+        Err(err) => CResult {
+            ok: None,
+            err: Some(err.to_string().into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_bytes_mixed_to_rln_proof(
+    bytes: &repr_c::Vec<u8>,
+) -> CResult<repr_c::Box<FFI_RLNProof>, repr_c::String> {
+    match <RLNProof as CanonicalDeserializeMixed>::deserialize(&bytes[..]) {
+        Ok(p) => CResult {
+            ok: Some(Box_::new(FFI_RLNProof(p))),
+            err: None,
+        },
+        Err(err) => CResult {
+            ok: None,
+            err: Some(err.to_string().into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_rln_proof_free(rln_proof: repr_c::Box<FFI_RLNProof>) {
+    drop(rln_proof);
+}
+
+// FFI_RLNPartialProof
+
+#[derive_ReprC]
+#[repr(opaque)]
+pub struct FFI_RLNPartialProof(pub(crate) PartialProof);
+
+#[ffi_export]
+pub fn ffi_rln_partial_proof_to_bytes_le(
+    partial_proof: &repr_c::Box<FFI_RLNPartialProof>,
+) -> CResult<repr_c::Vec<u8>, repr_c::String> {
+    let mut bytes = Vec::new();
+    match partial_proof.0.serialize_compressed(&mut bytes) {
+        Ok(()) => CResult {
+            ok: Some(bytes.into()),
+            err: None,
+        },
+        Err(err) => CResult {
+            ok: None,
+            err: Some(err.to_string().into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_bytes_le_to_rln_partial_proof(
+    bytes: &repr_c::Vec<u8>,
+) -> CResult<repr_c::Box<FFI_RLNPartialProof>, repr_c::String> {
+    match PartialProof::deserialize_compressed(&bytes[..]) {
+        Ok(p) => CResult {
+            ok: Some(Box_::new(FFI_RLNPartialProof(p))),
+            err: None,
+        },
+        Err(err) => CResult {
+            ok: None,
+            err: Some(err.to_string().into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_rln_partial_proof_free(partial_proof: repr_c::Box<FFI_RLNPartialProof>) {
+    drop(partial_proof);
+}
+
+// FFI_RLNProofValues
 
 #[derive_ReprC]
 #[repr(opaque)]
@@ -717,28 +1137,34 @@ pub struct FFI_RLNProofValues(pub(crate) RLNProofValues);
 
 #[ffi_export]
 pub fn ffi_rln_proof_values_get_root(pv: &repr_c::Box<FFI_RLNProofValues>) -> repr_c::Box<CFr> {
-    CFr::from(*pv.0.root()).into()
+    CFr::from(pv.0.root()).into()
 }
 
 #[ffi_export]
 pub fn ffi_rln_proof_values_get_x(pv: &repr_c::Box<FFI_RLNProofValues>) -> repr_c::Box<CFr> {
-    CFr::from(*pv.0.x()).into()
+    CFr::from(pv.0.x()).into()
 }
 
 #[ffi_export]
 pub fn ffi_rln_proof_values_get_external_nullifier(
     pv: &repr_c::Box<FFI_RLNProofValues>,
 ) -> repr_c::Box<CFr> {
-    CFr::from(*pv.0.external_nullifier()).into()
+    CFr::from(pv.0.external_nullifier()).into()
 }
 
 #[ffi_export]
 pub fn ffi_rln_proof_values_get_y(
     pv: &repr_c::Box<FFI_RLNProofValues>,
 ) -> CResult<repr_c::Box<CFr>, repr_c::String> {
-    CResult {
-        ok: Some(CFr::from(*pv.0.y()).into()),
-        err: None,
+    match pv.0.y() {
+        Some(y) => CResult {
+            ok: Some(CFr::from(y).into()),
+            err: None,
+        },
+        None => CResult {
+            ok: None,
+            err: Some("values are Multi; use get_ys".into()),
+        },
     }
 }
 
@@ -746,9 +1172,15 @@ pub fn ffi_rln_proof_values_get_y(
 pub fn ffi_rln_proof_values_get_nullifier(
     pv: &repr_c::Box<FFI_RLNProofValues>,
 ) -> CResult<repr_c::Box<CFr>, repr_c::String> {
-    CResult {
-        ok: Some(CFr::from(*pv.0.nullifier()).into()),
-        err: None,
+    match pv.0.nullifier() {
+        Some(n) => CResult {
+            ok: Some(CFr::from(n).into()),
+            err: None,
+        },
+        None => CResult {
+            ok: None,
+            err: Some("values are Multi; use get_nullifiers".into()),
+        },
     }
 }
 
@@ -756,9 +1188,15 @@ pub fn ffi_rln_proof_values_get_nullifier(
 pub fn ffi_rln_proof_values_get_selector_used(
     pv: &repr_c::Box<FFI_RLNProofValues>,
 ) -> CResult<repr_c::Vec<bool>, repr_c::String> {
-    CResult {
-        ok: Some(pv.0.selector_used().to_vec().into()),
-        err: None,
+    match pv.0.selector_used() {
+        Some(s) => CResult {
+            ok: Some(s.to_vec().into()),
+            err: None,
+        },
+        None => CResult {
+            ok: None,
+            err: Some("values are Single; selector_used is Multi-only".into()),
+        },
     }
 }
 
@@ -766,15 +1204,20 @@ pub fn ffi_rln_proof_values_get_selector_used(
 pub fn ffi_rln_proof_values_get_ys(
     pv: &repr_c::Box<FFI_RLNProofValues>,
 ) -> CResult<repr_c::Vec<CFr>, repr_c::String> {
-    CResult {
-        ok: Some(
-            pv.0.ys()
-                .iter()
-                .map(|fr| CFr::from(*fr))
-                .collect::<Vec<_>>()
-                .into(),
-        ),
-        err: None,
+    match pv.0.ys() {
+        Some(ys) => CResult {
+            ok: Some(
+                ys.iter()
+                    .map(|fr| CFr::from(*fr))
+                    .collect::<Vec<_>>()
+                    .into(),
+            ),
+            err: None,
+        },
+        None => CResult {
+            ok: None,
+            err: Some("values are Single; use get_y".into()),
+        },
     }
 }
 
@@ -782,39 +1225,63 @@ pub fn ffi_rln_proof_values_get_ys(
 pub fn ffi_rln_proof_values_get_nullifiers(
     pv: &repr_c::Box<FFI_RLNProofValues>,
 ) -> CResult<repr_c::Vec<CFr>, repr_c::String> {
-    CResult {
-        ok: Some(
-            pv.0.nullifiers()
-                .iter()
-                .map(|fr| CFr::from(*fr))
-                .collect::<Vec<_>>()
-                .into(),
-        ),
-        err: None,
+    match pv.0.nullifiers() {
+        Some(ns) => CResult {
+            ok: Some(
+                ns.iter()
+                    .map(|fr| CFr::from(*fr))
+                    .collect::<Vec<_>>()
+                    .into(),
+            ),
+            err: None,
+        },
+        None => CResult {
+            ok: None,
+            err: Some("values are Single; use get_nullifier".into()),
+        },
     }
 }
 
 #[ffi_export]
-pub fn ffi_rln_proof_values_get_version_byte(pv: &repr_c::Box<FFI_RLNProofValues>) -> u8 {
-    pv.0.version_byte()
+pub fn ffi_rln_proof_values_to_bytes_le(
+    pv: &repr_c::Box<FFI_RLNProofValues>,
+) -> CResult<repr_c::Vec<u8>, repr_c::String> {
+    let mut bytes = Vec::new();
+    match pv.0.serialize_compressed(&mut bytes) {
+        Ok(()) => CResult {
+            ok: Some(bytes.into()),
+            err: None,
+        },
+        Err(err) => CResult {
+            ok: None,
+            err: Some(err.to_string().into()),
+        },
+    }
 }
 
 #[ffi_export]
-pub fn ffi_rln_proof_values_to_bytes_le(pv: &repr_c::Box<FFI_RLNProofValues>) -> repr_c::Vec<u8> {
-    rln_proof_values_to_bytes_le(&pv.0).into()
-}
-
-#[ffi_export]
-pub fn ffi_rln_proof_values_to_bytes_be(pv: &repr_c::Box<FFI_RLNProofValues>) -> repr_c::Vec<u8> {
-    rln_proof_values_to_bytes_be(&pv.0).into()
+pub fn ffi_rln_proof_values_to_bytes_be(
+    pv: &repr_c::Box<FFI_RLNProofValues>,
+) -> CResult<repr_c::Vec<u8>, repr_c::String> {
+    let mut bytes = Vec::new();
+    match CanonicalSerializeBE::serialize(&pv.0, &mut bytes) {
+        Ok(()) => CResult {
+            ok: Some(bytes.into()),
+            err: None,
+        },
+        Err(err) => CResult {
+            ok: None,
+            err: Some(err.to_string().into()),
+        },
+    }
 }
 
 #[ffi_export]
 pub fn ffi_bytes_le_to_rln_proof_values(
     bytes: &repr_c::Vec<u8>,
 ) -> CResult<repr_c::Box<FFI_RLNProofValues>, repr_c::String> {
-    match bytes_le_to_rln_proof_values(bytes) {
-        Ok((pv, _)) => CResult {
+    match RLNProofValues::deserialize_compressed(&bytes[..]) {
+        Ok(pv) => CResult {
             ok: Some(Box_::new(FFI_RLNProofValues(pv))),
             err: None,
         },
@@ -829,8 +1296,8 @@ pub fn ffi_bytes_le_to_rln_proof_values(
 pub fn ffi_bytes_be_to_rln_proof_values(
     bytes: &repr_c::Vec<u8>,
 ) -> CResult<repr_c::Box<FFI_RLNProofValues>, repr_c::String> {
-    match bytes_be_to_rln_proof_values(bytes) {
-        Ok((pv, _)) => CResult {
+    match <RLNProofValues as CanonicalDeserializeBE>::deserialize(&bytes[..]) {
+        Ok(pv) => CResult {
             ok: Some(Box_::new(FFI_RLNProofValues(pv))),
             err: None,
         },
@@ -846,173 +1313,8 @@ pub fn ffi_rln_proof_values_free(proof_values: repr_c::Box<FFI_RLNProofValues>) 
     drop(proof_values);
 }
 
-// Proof generation APIs
-
 #[ffi_export]
-pub fn ffi_generate_rln_proof(
-    rln: &repr_c::Box<FFI_RLN>,
-    witness: &repr_c::Box<FFI_RLNWitnessInput>,
-) -> CResult<repr_c::Box<FFI_RLNProof>, repr_c::String> {
-    match rln.0.generate_rln_proof(&witness.0) {
-        Ok((proof, proof_values)) => {
-            let rln_proof = RLNProof {
-                proof_values,
-                proof,
-            };
-            CResult {
-                ok: Some(Box_::new(FFI_RLNProof(rln_proof))),
-                err: None,
-            }
-        }
-        Err(err) => CResult {
-            ok: None,
-            err: Some(err.to_string().into()),
-        },
-    }
-}
-
-#[ffi_export]
-pub fn ffi_generate_rln_proof_with_witness(
-    rln: &repr_c::Box<FFI_RLN>,
-    calculated_witness: &repr_c::Vec<repr_c::String>,
-    witness: &repr_c::Box<FFI_RLNWitnessInput>,
-) -> CResult<repr_c::Box<FFI_RLNProof>, repr_c::String> {
-    let calculated_witness_bigint: Result<Vec<BigInt>, _> = calculated_witness
-        .iter()
-        .map(|s| {
-            let s_str = unsafe { std::str::from_utf8_unchecked(s.as_bytes()) };
-            s_str.parse::<BigInt>()
-        })
-        .collect();
-
-    let calculated_witness_bigint = match calculated_witness_bigint {
-        Ok(w) => w,
-        Err(err) => {
-            return CResult {
-                ok: None,
-                err: Some(format!("Failed to parse witness: {}", err).into()),
-            }
-        }
-    };
-
-    match rln
-        .0
-        .generate_rln_proof_with_witness(calculated_witness_bigint, &witness.0)
-    {
-        Ok((proof, proof_values)) => {
-            let rln_proof = RLNProof {
-                proof_values,
-                proof,
-            };
-            CResult {
-                ok: Some(Box_::new(FFI_RLNProof(rln_proof))),
-                err: None,
-            }
-        }
-        Err(err) => CResult {
-            ok: None,
-            err: Some(err.to_string().into()),
-        },
-    }
-}
-
-// Partial proof generation APIs
-
-#[ffi_export]
-pub fn ffi_generate_partial_zk_proof(
-    rln: &repr_c::Box<FFI_RLN>,
-    partial_witness: &repr_c::Box<FFI_RLNPartialWitnessInput>,
-) -> CResult<repr_c::Box<FFI_RLNPartialProof>, repr_c::String> {
-    match rln.0.generate_partial_zk_proof(&partial_witness.0) {
-        Ok(partial_proof) => CResult {
-            ok: Some(Box_::new(FFI_RLNPartialProof(partial_proof))),
-            err: None,
-        },
-        Err(err) => CResult {
-            ok: None,
-            err: Some(err.to_string().into()),
-        },
-    }
-}
-
-#[ffi_export]
-pub fn ffi_finish_rln_proof(
-    rln: &repr_c::Box<FFI_RLN>,
-    partial_proof: &repr_c::Box<FFI_RLNPartialProof>,
-    witness: &repr_c::Box<FFI_RLNWitnessInput>,
-) -> CResult<repr_c::Box<FFI_RLNProof>, repr_c::String> {
-    match rln.0.finish_rln_proof(&partial_proof.0, &witness.0) {
-        Ok((proof, proof_values)) => {
-            let rln_proof = RLNProof {
-                proof_values,
-                proof,
-            };
-            CResult {
-                ok: Some(Box_::new(FFI_RLNProof(rln_proof))),
-                err: None,
-            }
-        }
-        Err(err) => CResult {
-            ok: None,
-            err: Some(err.to_string().into()),
-        },
-    }
-}
-
-// Proof verification APIs
-
-#[cfg(not(feature = "stateless"))]
-#[ffi_export]
-pub fn ffi_verify_rln_proof(
-    rln: &repr_c::Box<FFI_RLN>,
-    rln_proof: &repr_c::Box<FFI_RLNProof>,
-    x: &CFr,
-) -> CBoolResult {
-    match rln
-        .0
-        .verify_rln_proof(&rln_proof.0.proof, &rln_proof.0.proof_values, &x.0)
-    {
-        Ok(verified) => CBoolResult {
-            ok: verified,
-            err: None,
-        },
-        Err(err) => CBoolResult {
-            ok: false,
-            err: Some(err.to_string().into()),
-        },
-    }
-}
-
-#[ffi_export]
-pub fn ffi_verify_with_roots(
-    rln: &repr_c::Box<FFI_RLN>,
-    rln_proof: &repr_c::Box<FFI_RLNProof>,
-    roots: &repr_c::Vec<CFr>,
-    x: &CFr,
-) -> CBoolResult {
-    let roots_fr: Vec<Fr> = roots.iter().map(|cfr| cfr.0).collect();
-
-    match rln.0.verify_with_roots(
-        &rln_proof.0.proof,
-        &rln_proof.0.proof_values,
-        &x.0,
-        &roots_fr,
-    ) {
-        Ok(verified) => CBoolResult {
-            ok: verified,
-            err: None,
-        },
-        Err(err) => CBoolResult {
-            ok: false,
-            err: Some(err.to_string().into()),
-        },
-    }
-}
-
-// Identity secret recovery API
-
-#[ffi_export]
-pub fn ffi_compute_id_secret(
+pub fn ffi_rln_compute_id_secret(
     share1_x: &CFr,
     share1_y: &CFr,
     share2_x: &CFr,
@@ -1033,11 +1335,11 @@ pub fn ffi_compute_id_secret(
 }
 
 #[ffi_export]
-pub fn ffi_recover_id_secret(
+pub fn ffi_rln_recover_id_secret(
     proof_values_1: &repr_c::Box<FFI_RLNProofValues>,
     proof_values_2: &repr_c::Box<FFI_RLNProofValues>,
 ) -> CResult<repr_c::Box<CFr>, repr_c::String> {
-    match recover_id_secret(&proof_values_1.0, &proof_values_2.0) {
+    match proof_values_1.0.recover_secret(&proof_values_2.0) {
         Ok(secret) => CResult {
             ok: Some(Box_::new(CFr::from(*secret))),
             err: None,
@@ -1045,6 +1347,252 @@ pub fn ffi_recover_id_secret(
         Err(err) => CResult {
             ok: None,
             err: Some(err.to_string().into()),
+        },
+    }
+}
+
+// FFI_RLNMerkleProof
+
+#[derive_ReprC]
+#[repr(C)]
+pub struct FFI_RLNMerkleProof {
+    pub path_elements: repr_c::Vec<CFr>,
+    pub path_index: repr_c::Vec<u8>,
+}
+
+#[ffi_export]
+pub fn ffi_rln_merkle_proof_free(merkle_proof: repr_c::Box<FFI_RLNMerkleProof>) {
+    drop(merkle_proof);
+}
+
+#[ffi_export]
+pub fn ffi_rln_delete_leaf(rln: &mut repr_c::Box<FFI_RLN>, index: usize) -> CBoolResult {
+    match rln.0.delete_leaf(index) {
+        Ok(_) => CBoolResult {
+            ok: true,
+            err: None,
+        },
+        Err(err) => CBoolResult {
+            ok: false,
+            err: Some(err.into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_rln_set_leaf(rln: &mut repr_c::Box<FFI_RLN>, index: usize, leaf: &CFr) -> CBoolResult {
+    match rln.0.set_leaf(index, leaf.0) {
+        Ok(_) => CBoolResult {
+            ok: true,
+            err: None,
+        },
+        Err(err) => CBoolResult {
+            ok: false,
+            err: Some(err.into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_rln_get_leaf(
+    rln: &repr_c::Box<FFI_RLN>,
+    index: usize,
+) -> CResult<repr_c::Box<CFr>, repr_c::String> {
+    match rln.0.get_leaf(index) {
+        Ok(leaf) => CResult {
+            ok: Some(CFr::from(leaf).into()),
+            err: None,
+        },
+        Err(err) => CResult {
+            ok: None,
+            err: Some(err.into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_rln_leaves_set(rln: &repr_c::Box<FFI_RLN>) -> usize {
+    rln.0.leaves_set().unwrap_or(0)
+}
+
+#[ffi_export]
+pub fn ffi_rln_set_next_leaf(rln: &mut repr_c::Box<FFI_RLN>, leaf: &CFr) -> CBoolResult {
+    match rln.0.set_next_leaf(leaf.0) {
+        Ok(_) => CBoolResult {
+            ok: true,
+            err: None,
+        },
+        Err(err) => CBoolResult {
+            ok: false,
+            err: Some(err.into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_rln_set_leaves_from(
+    rln: &mut repr_c::Box<FFI_RLN>,
+    index: usize,
+    leaves: &repr_c::Vec<CFr>,
+) -> CBoolResult {
+    let leaves_vec: Vec<_> = leaves.iter().map(|cfr| cfr.0).collect();
+    match rln.0.set_leaves_from(index, leaves_vec) {
+        Ok(_) => CBoolResult {
+            ok: true,
+            err: None,
+        },
+        Err(err) => CBoolResult {
+            ok: false,
+            err: Some(err.into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_rln_init_tree_with_leaves(
+    rln: &mut repr_c::Box<FFI_RLN>,
+    leaves: &repr_c::Vec<CFr>,
+) -> CBoolResult {
+    let leaves_vec: Vec<_> = leaves.iter().map(|cfr| cfr.0).collect();
+    match rln.0.init_tree_with_leaves(leaves_vec) {
+        Ok(_) => CBoolResult {
+            ok: true,
+            err: None,
+        },
+        Err(err) => CBoolResult {
+            ok: false,
+            err: Some(err.into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_rln_atomic_operation(
+    rln: &mut repr_c::Box<FFI_RLN>,
+    index: usize,
+    leaves: &repr_c::Vec<CFr>,
+    indices: &repr_c::Vec<usize>,
+) -> CBoolResult {
+    let leaves_vec: Vec<_> = leaves.iter().map(|cfr| cfr.0).collect();
+    let indices_vec: Vec<_> = indices.iter().copied().collect();
+    match rln.0.atomic_operation(index, leaves_vec, indices_vec) {
+        Ok(_) => CBoolResult {
+            ok: true,
+            err: None,
+        },
+        Err(err) => CBoolResult {
+            ok: false,
+            err: Some(err.into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_rln_seq_atomic_operation(
+    rln: &mut repr_c::Box<FFI_RLN>,
+    leaves: &repr_c::Vec<CFr>,
+    indices: &repr_c::Vec<u8>,
+) -> CBoolResult {
+    let index = match rln.0.leaves_set() {
+        Ok(i) => i,
+        Err(err) => {
+            return CBoolResult {
+                ok: false,
+                err: Some(err.into()),
+            }
+        }
+    };
+    let leaves_vec: Vec<_> = leaves.iter().map(|cfr| cfr.0).collect();
+    let indices_vec: Vec<_> = indices.iter().map(|x| *x as usize).collect();
+    match rln.0.atomic_operation(index, leaves_vec, indices_vec) {
+        Ok(_) => CBoolResult {
+            ok: true,
+            err: None,
+        },
+        Err(err) => CBoolResult {
+            ok: false,
+            err: Some(err.into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_rln_get_root(rln: &repr_c::Box<FFI_RLN>) -> repr_c::Box<CFr> {
+    let root = rln.0.get_root().unwrap_or_else(|_| Fr::from(0u64));
+    CFr::from(root).into()
+}
+
+#[ffi_export]
+pub fn ffi_rln_get_merkle_proof(
+    rln: &repr_c::Box<FFI_RLN>,
+    index: usize,
+) -> CResult<repr_c::Box<FFI_RLNMerkleProof>, repr_c::String> {
+    match rln.0.get_merkle_proof(index) {
+        Ok((path_elements, path_index)) => {
+            let path_elements: repr_c::Vec<CFr> = path_elements
+                .iter()
+                .map(|fr| CFr::from(*fr))
+                .collect::<Vec<_>>()
+                .into();
+            let path_index: repr_c::Vec<u8> = path_index.into();
+            CResult {
+                ok: Some(Box_::new(FFI_RLNMerkleProof {
+                    path_elements,
+                    path_index,
+                })),
+                err: None,
+            }
+        }
+        Err(err) => CResult {
+            ok: None,
+            err: Some(err.into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_rln_set_metadata(
+    rln: &mut repr_c::Box<FFI_RLN>,
+    metadata: &repr_c::Vec<u8>,
+) -> CBoolResult {
+    match rln.0.set_metadata(metadata) {
+        Ok(_) => CBoolResult {
+            ok: true,
+            err: None,
+        },
+        Err(err) => CBoolResult {
+            ok: false,
+            err: Some(err.into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_rln_get_metadata(
+    rln: &repr_c::Box<FFI_RLN>,
+) -> CResult<repr_c::Vec<u8>, repr_c::String> {
+    match rln.0.get_metadata() {
+        Ok(metadata) => CResult {
+            ok: Some(metadata.into()),
+            err: None,
+        },
+        Err(err) => CResult {
+            ok: None,
+            err: Some(err.into()),
+        },
+    }
+}
+
+#[ffi_export]
+pub fn ffi_rln_flush(rln: &mut repr_c::Box<FFI_RLN>) -> CBoolResult {
+    match rln.0.flush() {
+        Ok(_) => CBoolResult {
+            ok: true,
+            err: None,
+        },
+        Err(err) => CBoolResult {
+            ok: false,
+            err: Some(err.into()),
         },
     }
 }
