+++ rln/src/protocol/mod.rs
@@ -1,7 +1,6 @@
-// This crate collects all the underlying primitives used to implement RLN
+// This module collects all the underlying primitives used to implement RLN
 
 mod keygen;
-mod mode;
 mod proof;
 mod serialize;
 mod slashing;
@@ -9,33 +8,14 @@ mod witness;
 mod zk;
 
 pub use keygen::{extended_keygen, extended_seeded_keygen, keygen, seeded_keygen};
-pub use mode::{MessageMode, Stateful, Stateless};
-pub use proof::{
-    bytes_be_to_rln_partial_proof, bytes_be_to_rln_proof, bytes_be_to_rln_proof_values,
-    bytes_le_to_rln_partial_proof, bytes_le_to_rln_proof, bytes_le_to_rln_proof_values,
-    calculated_witness_to_field_elements, generate_zk_proof_with_witness,
-    rln_partial_proof_to_bytes_be, rln_partial_proof_to_bytes_le, rln_proof_to_bytes_be,
-    rln_proof_to_bytes_le, rln_proof_values_to_bytes_be, rln_proof_values_to_bytes_le,
-    verify_zk_proof, RLNProof, RLNProofV3, RLNProofValues, RLNProofValuesMulti,
-    RLNProofValuesSingle, RLNProofValuesV3,
-};
-#[cfg(not(target_arch = "wasm32"))]
-pub use proof::{
-    finish_zk_proof, finish_zk_proof_with_rs, generate_partial_zk_proof, generate_zk_proof,
-    generate_zk_proof_with_rs,
-};
+pub use proof::{RLNProof, RLNProofValues, RLNProofValuesMulti, RLNProofValuesSingle};
 pub use serialize::{
     CanonicalDeserializeBE, CanonicalDeserializeMixed, CanonicalSerializeBE,
     CanonicalSerializeMixed, ENUM_TAG_MULTI, ENUM_TAG_SINGLE, ENUM_TAG_SIZE, FR_BYTE_SIZE,
     FR_LIMB_BYTE_SIZE, VEC_LEN_BYTE_SIZE,
 };
-pub use slashing::{compute_id_secret, recover_id_secret};
+pub use slashing::compute_id_secret;
 pub use witness::{
-    bytes_be_to_rln_partial_witness, bytes_be_to_rln_witness, bytes_le_to_rln_partial_witness,
-    bytes_le_to_rln_witness, compute_tree_root, proof_values_from_witness,
-    rln_partial_witness_to_bytes_be, rln_partial_witness_to_bytes_le, rln_witness_to_bigint_json,
-    rln_witness_to_bytes_be, rln_witness_to_bytes_le, RLNPartialWitnessInput,
-    RLNPartialWitnessInputV3, RLNWitnessInput, RLNWitnessInputMulti, RLNWitnessInputSingle,
-    RLNWitnessInputV3,
+    RLNPartialWitnessInput, RLNWitnessInput, RLNWitnessInputMulti, RLNWitnessInputSingle,
 };
 pub use zk::{RLNPartialZkProof, RLNZkProof, RecoverSecret};
