+++ rln/src/protocol/slashing.rs
@@ -1,14 +1,14 @@
 use ark_ff::AdditiveGroup;
 
-use super::proof::{RLNOutputs, RLNProofValues};
-use crate::{circuit::Fr, error::RecoverSecretError, utils::IdSecret};
+use crate::{
+    circuit::{Fr, IdSecret},
+    error::RecoverSecretError,
+};
 
 /// Computes identity secret from two (x, y) shares.
 ///
-/// This function is feature-agnostic: it operates on plain `(x, y)` share pairs
-/// that can be extracted from any [`RLNProofValues`] variant (`SingleV1` or `MultiV1`).
-/// By normalizing proof outputs into simple `(x, y)` pairs with a single nullifier,
-/// slashing can be performed across different prover modes without compile-time feature constraints.
+/// It operates on plain `(x, y)` share pairs that can be extracted from any
+/// proof values variant, so slashing can be performed across different modes.
 pub fn compute_id_secret(
     share1: (Fr, Fr),
     share2: (Fr, Fr),
@@ -32,69 +32,5 @@ pub fn compute_id_secret(
         Ok(id_secret)
     } else {
         Err(RecoverSecretError::DivisionByZero)
-    }
-}
-
-/// Recovers identity secret from two [`RLNProofValues`] with the same external nullifier.
-///
-/// This is a convenience API that accepts two proof values of the **same** variant.
-/// For cross-mode slashing (e.g. one share from `SingleV1` and another from `MultiV1`),
-/// extract the `(x, y)` pair and nullifier from each proof value and call [`compute_id_secret`] directly.
-pub fn recover_id_secret(
-    rln_proof_values_1: &RLNProofValues,
-    rln_proof_values_2: &RLNProofValues,
-) -> Result<IdSecret, RecoverSecretError> {
-    let external_nullifier_1 = rln_proof_values_1.external_nullifier();
-    let external_nullifier_2 = rln_proof_values_2.external_nullifier();
-
-    // We continue only if the proof values are for the same external nullifier
-    if external_nullifier_1 != external_nullifier_2 {
-        return Err(RecoverSecretError::ExternalNullifierMismatch(
-            *external_nullifier_1,
-            *external_nullifier_2,
-        ));
-    }
-
-    match (&rln_proof_values_1.outputs, &rln_proof_values_2.outputs) {
-        (RLNOutputs::SingleV1 { y: y1, .. }, RLNOutputs::SingleV1 { y: y2, .. }) => {
-            let share1 = (*rln_proof_values_1.x(), *y1);
-            let share2 = (*rln_proof_values_2.x(), *y2);
-            compute_id_secret(share1, share2)
-        }
-        (
-            RLNOutputs::MultiV1 {
-                ys: ys1,
-                nullifiers: nullifiers1,
-                selector_used: selector_used1,
-            },
-            RLNOutputs::MultiV1 {
-                ys: ys2,
-                nullifiers: nullifiers2,
-                selector_used: selector_used2,
-            },
-        ) => {
-            for (i, (nullifier_i, &used_i)) in
-                nullifiers1.iter().zip(selector_used1.iter()).enumerate()
-            {
-                if !used_i {
-                    continue;
-                }
-                for (j, (nullifier_j, &used_j)) in
-                    nullifiers2.iter().zip(selector_used2.iter()).enumerate()
-                {
-                    if !used_j {
-                        continue;
-                    }
-                    if nullifier_i == nullifier_j {
-                        let share1 = (*rln_proof_values_1.x(), ys1[i]);
-                        let share2 = (*rln_proof_values_2.x(), ys2[j]);
-                        return compute_id_secret(share1, share2);
-                    }
-                }
-            }
-            Err(RecoverSecretError::NoMatchingNullifier)
-        }
-        // Cross-mode slashing: extract (x, y) pairs and call compute_id_secret directly.
-        _ => Err(RecoverSecretError::NoMatchingNullifier),
     }
 }
